Get rid of all unnecessary C++ UI code (linux)
