+++ src/ui/linux/TogglDesktop/autocompletelistmodel.cpp
@@ -1,5 +1,7 @@
 #include "autocompletelistmodel.h"
+
 #include <QDebug>
+#include <QQmlEngine>
 
 AutocompleteListModel::AutocompleteListModel(QObject *parent, QVector<AutocompleteView*> list, uint64_t displayItem) :
     QAbstractListModel(parent),
@@ -68,4 +70,62 @@ QVariant AutocompleteListModel::data(const QModelIndex &index, int role) const {
     if (role == Qt::UserRole)
         return QVariant::fromValue(list.at(index.row()));
     return QVariant();
+}
+
+
+AutocompleteProxyModel::AutocompleteProxyModel(QObject *parent)
+    : QSortFilterProxyModel(parent) {
+    setFilterRole(Qt::UserRole);
+}
+
+int AutocompleteProxyModel::count() {
+    return rowCount(QModelIndex());
+}
+
+void AutocompleteProxyModel::setFilter(const QString &filter) {
+    setFilterRegExp(filter);
+}
+
+AutocompleteView *AutocompleteProxyModel::get(int idx) {
+    if (idx >= 0 && idx < rowCount()) {
+        auto d = data(index(idx, 0), Qt::UserRole);
+        if (!d.isValid())
+            return nullptr;
+        auto ptr = qvariant_cast<AutocompleteView*>(d);
+        if (!ptr)
+            return nullptr;
+        QQmlEngine::setObjectOwnership(ptr, QQmlEngine::CppOwnership);
+        return ptr;
+    }
+    return nullptr;
+}
+
+bool AutocompleteProxyModel::filterAcceptsRow(int source_row, const QModelIndex &source_parent) const {
+    QString input = filterRegExp().pattern();
+    QStringList words = input.split(" ");
+
+    auto variant = sourceModel()->data(sourceModel()->index(source_row, 0), Qt::UserRole);
+    if (!variant.isValid())
+        return false;
+    auto view = qvariant_cast<AutocompleteView*>(variant);
+    if (!view)
+        return false;
+    for (auto word : words) {
+        if (word.isEmpty() && words.count() > 1)
+            continue;
+        if (view->_Children.isEmpty()) {
+            if (view->Description.contains(word, Qt::CaseInsensitive))
+                    return true;
+            if (view->Text.contains(word, Qt::CaseInsensitive))
+                    return true;
+        }
+        for (auto v : view->_Children) {
+            if (v->Description.contains(word, Qt::CaseInsensitive))
+                    return true;
+            if (v->Text.contains(word, Qt::CaseInsensitive))
+                    return true;
+        }
+    }
+
+    return false;
 }
