--- src/ui/linux/TogglDesktop/overlaywidget.h
@@ -1,40 +1,0 @@
-#ifndef OVERLAYWIDGET_H
-#define OVERLAYWIDGET_H
-
-#include <QWidget>
-#include <QStackedWidget>
-#include <stdint.h>
-#include "./timeentryview.h"
-
-namespace Ui {
-class OverlayWidget;
-}
-
-class OverlayWidget : public QWidget
-{
-    Q_OBJECT
-
-public:
-    explicit OverlayWidget(QStackedWidget *parent = nullptr);
-    ~OverlayWidget();
-    void display();
-
-private:
-    Ui::OverlayWidget *ui;
-    int64_t current_type;
-
-private slots:  // NOLINT
-
-   void displayOverlay(const int64_t type);
-
-   void displayLogin(
-       const bool open,
-       const uint64_t user_id);
-
-   void on_actionButton_clicked();
-   void on_bottomText_linkActivated(const QString &link);
-   void on_topText_linkActivated(const QString &link);
-};
-
-
-#endif // OVERLAYWIDGET_H
